requirements.txt
fastapi
uvicorn[standard]
google-generativeai
Pillow
python-multipart